@@ -6,35 +6,48 @@
 name = cowinslot_tracker_app
 description = Add a short description here!
 author = filopd
-author-email = blackbeastfilo@live.com
-license = mit
-long-description = file: README.rst
-long-description-content-type = text/x-rst; charset=UTF-8
+author_email = blackbeastfilo@live.com
+license = MIT
+long_description = file: README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
 url = https://github.com/pyscaffold/pyscaffold/
-project-urls =
+# Add here related links, for example:
+project_urls =
     Documentation = https://pyscaffold.org/
+#    Source = https://github.com/pyscaffold/pyscaffold/
+#    Changelog = https://pyscaffold.org/en/latest/changelog.html
+#    Tracker = https://github.com/pyscaffold/pyscaffold/issues
+#    Conda-Forge = https://anaconda.org/conda-forge/pyscaffold
+#    Download = https://pypi.org/project/PyScaffold/#files
+#    Twitter = https://twitter.com/PyScaffold
+
 # Change if running only on Windows, Mac or Linux (comma-separated)
 platforms = any
+
 # Add here all kinds of additional classifiers as defined under
 # https://pypi.python.org/pypi?%3Aaction=list_classifiers
 classifiers =
     Development Status :: 4 - Beta
     Programming Language :: Python
 
+
 [options]
 zip_safe = False
-packages = find:
+packages = find_namespace:
 include_package_data = True
 package_dir =
     =src
-# DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
-setup_requires = pyscaffold>=4.0a0,<4.1a0
-# Add here dependencies of your project (semicolon/line-separated), e.g.
-# install_requires = numpy; scipy
-# The usage of test_requires is discouraged, see `Dependency Management` docs
-# tests_require = pytest; pytest-cov
-# Require a specific Python version, e.g. Python 2.7 or >= 3.4
-# python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+
+# Require a min/specific Python version (comma-separated conditions)
+# python_requires = >=3.8
+
+# Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
+# Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
+# new major versions. This works if the required packages follow Semantic Versioning.
+# For more information, check out https://semver.org/.
+install_requires =
+    importlib-metadata; python_version<"3.8"
+
 
 [options.packages.find]
 where = src
