@@ -1,10 +1,10 @@
-# -*- coding: utf-8 -*-
 """
     Dummy conftest.py for cowinslot_tracker_app.
 
     If you don't know what this is for, just leave it empty.
     Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
+    - https://docs.pytest.org/en/stable/fixture.html
+    - https://docs.pytest.org/en/stable/writing_plugins.html
 """
 
 # import pytest
