@@ -1,16 +1,25 @@
-# -*- coding: utf-8 -*-
-
 import pytest
-from cowinslot_tracker_app.skeleton import fib
+
+from cowinslot_tracker_app.skeleton import fib, main
 
 __author__ = "filopd"
 __copyright__ = "filopd"
-__license__ = "mit"
+__license__ = "MIT"
 
 
 def test_fib():
+    """API Tests"""
     assert fib(1) == 1
     assert fib(2) == 1
     assert fib(7) == 13
     with pytest.raises(AssertionError):
         fib(-10)
+
+
+def test_main(capsys):
+    """CLI Tests"""
+    # capsys is a pytest fixture that allows asserts agains stdout/stderr
+    # https://docs.pytest.org/en/stable/capture.html
+    main(["7"])
+    captured = capsys.readouterr()
+    assert "The 7-th Fibonacci number is 13" in captured.out
